requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==2.3.1
idna==3.1
multidict==5.1.0
typing-extensions==4.7.1
yarl==1.6.3
